Implemented spinner for Add Category Fragment.
NOTE: I had a few errors that I thought was related to the old layout file, so I deleted the old one and created a new one with copy pasted everything. Found out it was an issue related with how i created the spinner adapter. So that is why the layout file is deleted and a new one is being pushed.